@@ -33,20 +33,20 @@
     "json-loader": "^0.5.4",
     "jsx-control-statements": "3.1.2",
     "leaflet": "0.7.3",
-    "linkify-it": "^1.2.0",
-    "markdown-it": "^6.0.1",
+    "linkify-it": "^2.0.0",
+    "markdown-it": "^7.0.0",
     "mustache": "^2.2.1",
     "mutationobserver-shim": "^0.3.1",
     "proj4": "^2.3.14",
     "raw-loader": "^0.5.1",
-    "react-addons-pure-render-mixin": "0.14.8",
+    "react-addons-pure-render-mixin": "15.2.0",
     "resolve": "^1.1.7",
     "resolve-url-loader": "^1.4.3",
     "sass-loader": "^3.2.0",
-    "simple-statistics": "^1.0.1",
+    "simple-statistics": "^2.0.0",
     "string-replace-webpack-plugin": "0.0.3",
     "style-loader": "^0.13.1",
-    "superagent": "~1.7.0",
+    "superagent": "^2.0.0",
     "terriajs-cesium": "1.20.0",
     "togeojson": "^0.13.0",
     "urijs": "^1.17.1",
@@ -59,7 +59,7 @@
     "babel-eslint": "^6.0.2",
     "css-loader": "^0.23.1",
     "electron-prebuilt": "^1.2.3",
-    "eslint": "^2.7.0",
+    "eslint": "^3.0.0",
     "eslint-plugin-jsx-control-statements": "2.1.1",
     "eslint-plugin-react": "^4.3.0",
     "generate-terriajs-schema": "^1.2.1",
@@ -68,34 +68,34 @@
     "gulp-util": "^3.0.7",
     "jasmine-core": "^2.4.1",
     "jsdoc": "^3.4.0",
-    "karma": "^0.13.22",
-    "karma-browserstack-launcher": "^0.1.10",
-    "karma-chrome-launcher": "^0.2.3",
+    "karma": "^1.1.0",
+    "karma-browserstack-launcher": "^1.0.1",
+    "karma-chrome-launcher": "^1.0.1",
     "karma-detect-browsers": "^2.0.2",
     "karma-electron-launcher": "^0.1.0",
-    "karma-firefox-launcher": "^0.1.7",
-    "karma-ie-launcher": "^0.2.0",
-    "karma-jasmine": "^0.3.8",
-    "karma-opera-launcher": "^0.3.0",
-    "karma-safari-launcher": "^0.1.1",
-    "karma-sauce-launcher": "^0.3.1",
-    "karma-spec-reporter": "^0.0.25",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-ie-launcher": "^1.0.0",
+    "karma-jasmine": "^1.0.2",
+    "karma-opera-launcher": "^1.0.0",
+    "karma-safari-launcher": "^1.0.0",
+    "karma-sauce-launcher": "^1.0.0",
+    "karma-spec-reporter": "^0.0.26",
     "kss": "^2.1.1",
     "node-notifier": "^4.5.0",
     "node-sass": "^3.4.2",
-    "react": "0.14.x",
-    "react-addons-test-utils": "0.14.7",
-    "react-dom": "0.14.x",
-    "react-shallow-testutils": "^1.0.0",
+    "react": "15.2.0",
+    "react-addons-test-utils": "15.2.0",
+    "react-dom": "15.2.0",
+    "react-shallow-testutils": "^2.0.0",
     "svg-sprite-loader": "0.0.23",
     "socket.io": "1.4.7",
     "terriajs-jasmine-ajax": "^3.2.1",
-    "terriajs-server": "^1.4.1",
+    "terriajs-server": "^2.3.0",
     "webpack-dev-server": "^1.14.1"
   },
   "peerDependencies": {
-    "react": "0.14.x",
-    "react-dom": "0.14.x"
+    "react": "^15.2.0",
+    "react-dom": "^15.2.0"
   },
   "scripts": {
     "prepublish": "gulp post-npm-install",
