@@ -65,33 +65,26 @@ function CorsProxy(overrideLoadJson) {
 }
 
 /**
- * Initialises values with config from the proxy server. This is the recommended way to use this object as it ensures
+ * Initialises values with config previously loaded from server. This is the recommended way to use this object as it ensures
  * the options will be correct for the proxy server it's configured to call, but this can be skipped and the values it
- * initialises set manually if desired. Note that if the request to get config from the proxy server fails, an error
- * is logged and nothing else happens. If you want to do something more like display a warning to the user, you'll have
- * to .otherwise the promise returned from the method.
+ * initialises set manually if desired. 
  *
- * @param {String} proxyableDomainsUrl The url to call on the terriajs server in order to get proxyable domains and other config
+ * @param {Object} serverConfig Configuration options retrieved from a ServerConfig object.
  * @param {String} baseProxyUrl The base URL to proxy with - this will default to 'proxy/'
  * @param {String[]} proxyDomains Initial value for proxyDomains to which proxyable domains from the server will be appended -
  *      defaults to an empty array.
  * @returns {Promise} A promise that resolves when initialisation is complete.
  */
-CorsProxy.prototype.init = function(proxyableDomainsUrl, baseProxyUrl, proxyDomains) {
+CorsProxy.prototype.init = function(serverConfig, baseProxyUrl, proxyDomains) {
     this.baseProxyUrl = defaultValue(baseProxyUrl, DEFAULT_BASE_PROXY_PATH);
     this.proxyDomains = defaultValue(proxyDomains, []);
-
-    return this.loadJson(proxyableDomainsUrl).then(function(proxyableDomains) {
-        if (proxyableDomains.proxyAllDomains) {
-            this.isOpenProxy = true;
+    if (serverConfig && typeof serverConfig === 'object') {
+        this.isOpenProxy = !! serverConfig.proxyAllDomains;
+        // ignore client list of allowed proxies in favour of definitive server list.
+        if (Array.isArray(serverConfig.allowProxyFor)) {
+            this.proxyDomains = serverConfig.allowProxyFor;
         }
-        this.proxyDomains = this.proxyDomains.concat(proxyableDomains.proxyableDomains);
-    }.bind(this)).otherwise(function(e) {
-        console.error('Failed to resolve proxyableDomains from ' + proxyableDomainsUrl);
-
-        // Keep on throwing so that the caller of init can .otherwise the exception.
-        throw e;
-    });
+    }
 };
 
 /**
