@@ -105,6 +105,13 @@ var CkanCatalogGroup = function(terria) {
     this.groupBy = 'group';
 
     /**
+     * Gets or sets a title for the group holding all items that don't have a group in CKAN. If the value is a blank string or undefined,
+     * these items will be left at the top level, not grouped.
+     * @type {String}
+     */
+    this.ungroupedTitle = 'No group';
+
+    /**
      * Gets or sets a value indicating whether each catalog item's name should be populated from
      * individual resources instead of from the CKAN dataset.
      * @type {Boolean}
@@ -164,7 +171,7 @@ var CkanCatalogGroup = function(terria) {
      * Gets or sets a regular expression that, when it matches a resource's format, indicates that the resource is a KML resource.
      * @type {RegExp}
      */
-    this.kmlResourceFormat = /^kml$/i;
+    this.kmlResourceFormat = /^km[lz]$/i;
 
     /**
      * True to allow CSV resources to be added to the catalog; otherwise, false.
@@ -231,6 +238,7 @@ var CkanCatalogGroup = function(terria) {
         'blacklist',
         'wmsParameters',
         'groupBy',
+        'ungroupedTitle',
         'useResourceName',
         'allowEntireWmsServers',
         'allowEntireWfsServers',
@@ -350,7 +358,9 @@ CkanCatalogGroup.prototype._getValuesThatInfluenceLoad = function() {
         this.includeCsv,
         this.includeEsriMapServer,
         this.includeGeoJson,
-        this.includeCzml
+        this.includeCzml,
+        this.groupBy,
+        this.ungroupedTitle
     ];
 };
 
@@ -527,6 +537,8 @@ function createItemFromResource(resource, ckanGroup, itemData, extras, parent) {
     });
 }
 function populateGroupFromResults(ckanGroup, json) {
+    var ungrouped;
+
     var items = json.result.results;
     for (var itemIndex = 0; itemIndex < items.length; ++itemIndex) {
         var item = items[itemIndex];
@@ -581,7 +593,17 @@ function populateGroupFromResults(ckanGroup, json) {
                     }
                 }
             } else {
-                addItem(resource, ckanGroup, item, extras, ckanGroup);
+                if (!ckanGroup.ungroupedTitle) {
+                    addItem(resource, ckanGroup, item, extras, ckanGroup);
+                } else {
+                    if (!defined(ungrouped)) {
+                        ungrouped = new CatalogGroup(ckanGroup.terria);
+                        ungrouped.name = ckanGroup.ungroupedTitle;
+                        ungrouped.id = ckanGroup.uniqueId + '/_ungrouped';
+                        ckanGroup.add(ungrouped);
+                    }
+                    addItem(resource, ckanGroup, item, extras, ungrouped);
+                }
             }
         }
     }
