@@ -22,6 +22,7 @@ var CameraView = require('./CameraView');
 var Catalog = require('./Catalog');
 var ConsoleAnalytics = require('../Core/ConsoleAnalytics');
 var CorsProxy = require('../Core/CorsProxy');
+var ServerConfig = require('../Core/ServerConfig');
 var GoogleAnalytics = require('../Core/GoogleAnalytics');
 var NowViewing = require('./NowViewing');
 var Services = require('./Services');
@@ -302,6 +303,19 @@ var Terria = function(options) {
     this.urlShortener = undefined;
 
     /**
+     * Gets or sets the shareDataService to be used with Terria, which can save JSON or (in future) other user-provided
+     * data somewhere. It can be used to generate short URLs.
+     * @type {Object}
+     */
+    this.shareDataService = undefined;
+
+    /**
+     * Gets or sets the ServerConfig object representing server-side configuration.
+     * @type {Object}
+     */
+    this.serverConfig = undefined;
+
+    /**
      * Event that tracks changes to the progress in loading new tiles from either Cesium or Leaflet - events will be
      * raised with the number of tiles that still need to load.
      *
@@ -367,6 +381,8 @@ Terria.prototype.start = function(options) {
 
     var applicationUrl = defaultValue(options.applicationUrl, '');
     this.urlShortener = options.urlShortener;
+    this.shareDataService = options.shareDataService;
+
 
     var that = this;
     return loadJson5(options.configUrl).then(function(config) {
@@ -392,24 +408,49 @@ Terria.prototype.start = function(options) {
                 that.initSources.push(generateInitializationUrl(initializationUrls[i]));
             }
         }
-
-        // Init the proxy THEN update from application URL as updating the application URL might have to expand
-        // a shortened URL which requires AJAX which might involved cross-origin AJAX which might require the
-        // list of proxyable domains to be loaded.
-        return that.corsProxy.init(cp.proxyableDomainsUrl, cp.corsProxyBaseUrl, config.proxyDomains)
-            .then(function () {
-                return that.updateApplicationUrl(applicationUrl, that.urlShortener).then(function () {
-                    if (options.defaultTo2D && !defined(that.userProperties.map)) {
-                        that.viewerMode = ViewerMode.Leaflet;
-                    }
-                    that.catalog.isLoading = false;
-                });
-            }).otherwise(function() {
-                that.error.raiseEvent({
-                    title: 'Failed to initialize proxy',
-                    message: 'This will cause some layers to be unavailable, or all layers on some older browsers.'
-                });
+        
+        that.serverConfig = new ServerConfig();
+        let serverConfig;
+        return that.serverConfig.init().then(function() {
+            // All the "proxyableDomains" bits here are due to a pre-serverConfig mechanism for whitelisting domains.
+            // We should deprecate it.
+            var pdu = that.configParameters.proxyableDomainsUrl;
+            if (pdu) {
+                return loadJson5(pdu);
+            }
+        }).then(function(proxyableDomains) {
+            if (proxyableDomains) {
+                // format of proxyableDomains JSON file slightly differs from serverConfig format. 
+                proxyableDomains.allowProxyFor = proxyableDomains.allowProxyFor || proxyableDomains.proxyableDomains;
+            }
+            if (typeof that.serverConfig === 'object') {
+                serverConfig = that.serverConfig.config; // if server config is unavailable, this remains undefined.
+            }
+            if (that.shareDataService) { 
+                that.shareDataService.init(serverConfig);
+            }
+            that.corsProxy.init(proxyableDomains || serverConfig, cp.corsProxyBaseUrl, config.proxyDomains);
+        }).otherwise(function(e) {
+            console.error(e);
+            // There's no particular reason an error should be thrown here.
+            that.error.raiseEvent({
+                title: 'Failed to initialize services',
+                message: 'A problem occurred with the Terria server. This may cause some layers or the conversion service to be unavailable.'
+            });            
+        }).then(function() {
+            return that.updateApplicationUrl(applicationUrl, that.urlShortener);
+        }).then(function () {
+            if (options.defaultTo2D && !defined(that.userProperties.map)) {
+                that.viewerMode = ViewerMode.Leaflet;
+            }
+            that.catalog.isLoading = false;
+        }).otherwise(function(e) {
+            console.error('Error from updateApplicationUrl: ',  e);
+            that.error.raiseEvent({
+                title: 'Problem loading URL',
+                message: 'A problem occurred while initialising Terria with URL parameters.'
             });
+        });
     });
 };
 
@@ -433,7 +474,6 @@ Terria.prototype.updateApplicationUrl = function(newUrl) {
         } else if (that.userProperties.map === '3d') {
             that.viewerMode = ViewerMode.CesiumTerrain;
         }
-
         return loadInitSources(that, initSources);
     });
 };
@@ -615,36 +655,40 @@ var latestStartVersion = '0.0.05';
 
 function interpretHash(terria, hashProperties, userProperties, persistentInitSources, temporaryInitSources) {
     var promise;
-    if (defined(terria.urlShortener) && defined(hashProperties.share)) {
-        promise = terria.urlShortener.expand(hashProperties.share);
+    // #share=xyz . Resolve with either share data service or URL shortener.
+    if (defined(hashProperties.share)) {
+        if (defined(terria.shareDataService)) {
+            // get JSON directly
+            promise = terria.shareDataService.resolveData(hashProperties.share);
+        } else if (defined(terria.urlShortener)) {
+            promise = terria.urlShortener.expand(hashProperties.share)
+                // get URL, and extract the JSON part
+                .then(longUrl => longUrl && queryToObject(new URI(longUrl).fragment()));
+        }
     }
 
-    return when(promise, function(longUrl) {
-        if (defined(longUrl)) {
-            var uri = new URI(longUrl);
-            var hash = uri.fragment();
-            hashProperties = combine(hashProperties, queryToObject(hash));
-        }
+    return when(promise, function(shareProps) {
+        Object.keys(hashProperties).forEach(function(property) {
+            var propertyValue = hashProperties[property];
 
-        for (var property in hashProperties) {
-            if (hashProperties.hasOwnProperty(property)) {
-                var propertyValue = hashProperties[property];
-
-                if (property === 'clean') {
-                    persistentInitSources.length = 0;
-                    temporaryInitSources.length = 0;
-                } else if (property === 'start') {
-                    var startData = JSON.parse(propertyValue);
-                    interpretStartData(startData, persistentInitSources, temporaryInitSources);
-                } else if (defined(propertyValue) && propertyValue.length > 0) {
-                    userProperties[property] = propertyValue;
-                    knockout.track(userProperties, [property]);
-                } else {
-                    var initSourceFile = generateInitializationUrl(property);
-                    persistentInitSources.push(initSourceFile);
-                    temporaryInitSources.push(initSourceFile);
-                }
+            if (property === 'clean') {
+                persistentInitSources.length = 0;
+                temporaryInitSources.length = 0;
+            } else if (property === 'start') {
+                // a share link that hasn't been shortened: JSON embedded in URL (only works for small quantities of JSON)  
+                var startData = JSON.parse(propertyValue);
+                interpretStartData(startData, persistentInitSources, temporaryInitSources);
+            } else if (defined(propertyValue) && propertyValue.length > 0) {
+                userProperties[property] = propertyValue;
+                knockout.track(userProperties, [property]);
+            } else {
+                var initSourceFile = generateInitializationUrl(property);
+                persistentInitSources.push(initSourceFile);
+                temporaryInitSources.push(initSourceFile);
             }
+        });
+        if (shareProps) {
+            interpretStartData(shareProps, persistentInitSources, temporaryInitSources);
         }
     });
 }
