@@ -3,7 +3,6 @@
 /*global require*/
 var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
-var PopupMessageConfirmationViewModel = require('../ViewModels/PopupMessageConfirmationViewModel.js');
 
 var createCatalogItemFromUrl = require('./createCatalogItemFromUrl');
 var createCatalogMemberFromType = require('./createCatalogMemberFromType');
@@ -28,23 +27,23 @@ var OgrCatalogItem = require('../Models/OgrCatalogItem');
  *                                           will be rejected with the string 'The user declined to use the conversion service.'.
  * @return {Promise} A promise that resolves to the created catalog item.
  */
-var createCatalogItemFromFileOrUrl = function(terria, fileOrUrl, dataType, confirmConversion) {
+var createCatalogItemFromFileOrUrl = function(terria, viewState, fileOrUrl, dataType, confirmConversion) {
     function tryConversionService(newItem) {
        if (name.match(/\.(shp|jpg|jpeg|pdf|xlsx|xls|tif|tiff|png|txt|doc|docx|xml|json)$/)) {
             terria.error.raiseEvent(new TerriaError({
                 title: 'Unsupported file type',
-                message: 'This file format is not supported by ' + terria.appName + '. Directly supported file formats include: ' + 
-                '<ul><li>.geojson</li><li>.kml, .kmz</li><li>.csv (in <a href="https://github.com/NICTA/nationalmap/wiki/csv-geo-au">csv-geo-au format</a>)</li></ul>' + 
+                message: 'This file format is not supported by ' + terria.appName + '. Directly supported file formats include: ' +
+                '<ul><li>.geojson</li><li>.kml, .kmz</li><li>.csv (in <a href="https://github.com/NICTA/nationalmap/wiki/csv-geo-au">csv-geo-au format</a>)</li></ul>' +
                 'File formats supported through the online conversion service include: ' +
                 '<ul><li>Shapefile (.zip)</li><li>MapInfo TAB (.zip)</li><li>Possibly other <a href="http://www.gdal.org/ogr_formats.html">OGR Vector Formats</a></li></ul>'
             }));
             return undefined;
-        } 
-        return getConfirmation(confirmConversion, 'This file is not directly supported by '+terria.appName+'.\n\n' +
+        }
+        return getConfirmation(terria, viewState, confirmConversion, 'This file is not directly supported by '+terria.appName+'.\n\n' +
             'Do you want to try uploading it to the '+terria.appName+' conversion service? This may work for ' +
             'small, zipped Esri Shapefiles or MapInfo TAB files.')
-            .then(function(confirmed) { 
-                return confirmed ? loadItem(new OgrCatalogItem(terria), name, fileOrUrl) : undefined; 
+            .then(function(confirmed) {
+                return confirmed ? loadItem(new OgrCatalogItem(terria), name, fileOrUrl) : undefined;
             });
     }
 
@@ -60,13 +59,13 @@ var createCatalogItemFromFileOrUrl = function(terria, fileOrUrl, dataType, confi
             } else {
                 // It's a file or service we support directly
                 // In some cases (web services), the item will already have been loaded by this point.
-                return loadItem(newItem, name, fileOrUrl); 
+                return loadItem(newItem, name, fileOrUrl);
             }
         });
     } else if (dataType === 'other') { // user explicitly chose "Other (use conversion service)"
-        return getConfirmation(confirmConversion, 'Ready to upload your file to the ' + terria.appName + ' conversion service?')
-            .then(function(confirmed) { 
-                return confirmed ? loadItem(createCatalogMemberFromType('ogr', terria), name, fileOrUrl) : undefined; 
+        return getConfirmation(terria, viewState, confirmConversion, 'Ready to upload your file to the ' + terria.appName + ' conversion service?')
+            .then(function(confirmed) {
+                return confirmed ? loadItem(createCatalogMemberFromType('ogr', terria), name, fileOrUrl) : undefined;
             });
     } else {
         // User has provided a type, so we go with that.
@@ -75,19 +74,19 @@ var createCatalogItemFromFileOrUrl = function(terria, fileOrUrl, dataType, confi
 };
 
 /* Returns a promise that returns true if user confirms, or false if they abort. */
-function getConfirmation(confirmConversion, message) {
+function getConfirmation(terria, viewState, confirmConversion, message) {
     if (!confirmConversion) {
         return when(true);
     }
-    
+
     var d = when.defer(); // there's no `when.promise(resolver)` in when 1.7.1
-    PopupMessageConfirmationViewModel.open('ui', { 
-        confirmText: 'Upload', 
+    viewState.notifications.push({
+        confirmText: 'Upload',
+        denyText: 'Cancel',
         title: 'Use conversion service?',
         message: message,
         confirmAction: function () { d.resolve(true); },
-        enableDeny: true,
-        denyAction: function() { this.close(); d.resolve(false); }
+        denyAction: function() { d.resolve(false); }
     });
     return d.promise;
 }
