@@ -16,19 +16,38 @@ const userPropWhiteList = ['hideExplorerPanel', 'activeTabId'];
  * @param terria The terria instance to serialize.
  * @returns {String} A URI that will rebuild the current state when viewed in a browser.
  */
-export function buildShareLink(terria) {
-    // Build the share URL.
-    const request = {
+export function buildShareLink(terria) { 
+    const uri = new URI(window.location)
+        .fragment('')
+        .search({ 'start': JSON.stringify(getShareData(terria)) });
+    
+    userPropWhiteList.forEach(key => uri.addSearch({ key: terria.userProperties[key] }));
+    return uri.fragment(uri.query()).query('').toString(); // replace ? with #
+}
+
+/**
+ * Returns just the JSON that defines the current view.
+ * @param  {Object} terria The Terria object.
+ * @return {Object}        
+ */
+function getShareData(terria) {
+    const initSources = terria.initSources.slice();
+
+    addUserAddedCatalog(terria, initSources);
+    addSharedMembers(terria, initSources);
+    addViewSettings(terria, initSources);
+    return {
         version: '0.0.05',
-        initSources: terria.initSources.slice()
+        initSources: initSources 
     };
-
-    addUserAddedCatalog(terria, request.initSources);
-    addSharedMembers(terria, request.initSources);
-    addViewSettings(terria, request.initSources);
-
-    const requestString = JSON.stringify(request);
-    return `${getUri()}#start=${encodeURIComponent(requestString)}${generateUserPropertiesQuery(terria)}`;
+}
+/**
+ * Is it currently possible to generate short URLs?
+ * @param  {Object} terria The Terria object.
+ * @return {Boolean}        
+ */
+export function canShorten(terria) {
+    return (terria.urlShortener && terria.urlShortener.isUsable) || (terria.shareDataService && terria.shareDataService.isUsable);
 }
 
 /**
@@ -37,20 +56,12 @@ export function buildShareLink(terria) {
  * @returns {Promise<String>} A promise that will return the shortened url when complete.
  */
 export function buildShortShareLink(terria) {
-    return terria.urlShortener.shorten(buildShareLink(terria)).then(function(token) {
-        return `${getUri()}#share=${token}`;
-    });
-}
-
-/**
- * Gets the browser's current URI.
- */
-function getUri() {
-    const uri = new URI(window.location);
-
-    // Remove the portion of the URL after the hash.
-    uri.fragment('');
-    return uri.toString();
+    const urlFromToken = token => new URI(window.location).fragment('share=' + token).toString(); 
+    if (defined(terria.shareDataService)) {
+        return terria.shareDataService.getShareToken(getShareData(terria)).then(urlFromToken);
+    } else {
+        return terria.urlShortener.shorten(buildShareLink(terria)).then(urlFromToken);
+    } // we assume that URL shortener is defined.
 }
 
 /**
@@ -161,19 +172,6 @@ function addViewSettings(terria, initSources) {
         baseMapName: terria.baseMap.name,
         viewerMode: terria.leaflet ? '2d' : '3d'
     });
-}
-
-/**
- * Generates a query string for custom user properties.
- *
- * @returns {String}
- * @private
- */
-function generateUserPropertiesQuery(terria) {
-    return userPropWhiteList.reduce((querySoFar, key) => {
-        const val = terria.userProperties[key];
-        return val ? `${querySoFar}&${key}=${encodeURIComponent(val)}` : querySoFar;
-    }, '');
 }
 
 /**
