@@ -1,13 +1,13 @@
-'use strict';
-
+import React from 'react';
 import classNames from 'classnames';
+
 import DataCatalogMember from '../DataCatalog/DataCatalogMember.jsx';
 import DataPreview from '../Preview/DataPreview.jsx';
 import defined from 'terriajs-cesium/Source/Core/defined';
 import MobileSearch from './MobileSearch.jsx';
 import WorkbenchList from '../Workbench/WorkbenchList.jsx';
 import ObserveModelMixin from '../ObserveModelMixin';
-import React from 'react';
+import Icon from '../Icon';
 
 import Styles from './mobile-modal-window.scss';
 
@@ -20,39 +20,39 @@ const MobileModalWindow = React.createClass({
     },
 
     renderModalContent() {
-        switch(this.props.viewState.mobileView) {
-        case this.props.viewState.mobileViewOptions.search:
-            return <div className={Styles.modalBg}>
-                    <MobileSearch terria={this.props.terria}
-                                  viewState={this.props.viewState}
-                   /></div>;
-        case this.props.viewState.mobileViewOptions.data:
-            return <div className={Styles.modalBg}>
+        switch (this.props.viewState.mobileView) {
+            case this.props.viewState.mobileViewOptions.search:
+                return (
+                    <MobileSearch terria={this.props.terria} viewState={this.props.viewState}/>
+                );
+            case this.props.viewState.mobileViewOptions.data:
+                return (
                     <ul className={Styles.dataCatalog}>
-                    {this.props.terria.catalog.group.items.filter(defined)
-                          .map((item, i) => (
-                            <DataCatalogMember viewState={this.props.viewState}
-                                               member={item}
-                                               key={item.uniqueId}
-                            />
-                    ))}
-                </ul>
-            </div>;
-        case this.props.viewState.mobileViewOptions.preview:
-            return <div className={Styles.modalBg}>
-                        <DataPreview terria={this.props.terria}
-                                     viewState={this.props.viewState}
-                                     previewed={this.props.viewState.previewedItem}
-                        />
-                    </div>;
-        case this.props.viewState.mobileViewOptions.nowViewing:
-            return <div className={Styles.modalBg}>
-                        <WorkbenchList viewState={this.props.viewState}
-                                       terria={this.props.terria}
-                        />
-                    </div>;
-        default:
-            return null;
+                        {this.props.terria.catalog.group.items.filter(defined)
+                            .map((item, i) => (
+                                <DataCatalogMember viewState={this.props.viewState}
+                                                   member={item}
+                                                   key={item.uniqueId}
+                                />
+                            ))
+                        }
+                    </ul>
+                );
+            case this.props.viewState.mobileViewOptions.preview:
+                return (
+                    <DataPreview terria={this.props.terria}
+                                 viewState={this.props.viewState}
+                                 previewed={this.props.viewState.previewedItem}
+                    />
+                );
+            case this.props.viewState.mobileViewOptions.nowViewing:
+                return (
+                    <WorkbenchList viewState={this.props.viewState}
+                                   terria={this.props.terria}
+                    />
+                );
+            default:
+                return null;
         }
     },
 
@@ -62,21 +62,46 @@ const MobileModalWindow = React.createClass({
     },
 
     componentWillReceiveProps() {
-        if((this.props.terria.nowViewing.items.length === 0) &&
-          (this.props.viewState.mobileView === this.props.viewState.mobileViewOptions.nowViewing)) {
+        if ((this.props.terria.nowViewing.items.length === 0) &&
+            (this.props.viewState.mobileView === this.props.viewState.mobileViewOptions.nowViewing)) {
             this.props.viewState.switchMobileView(null);
             this.props.viewState.explorerPanelIsVisible = false;
         }
     },
 
+    goBack() {
+        this.props.viewState.mobileView = this.props.viewState.mobileViewOptions.data;
+    },
+
     render() {
-        let modalClass = classNames(Styles.mobileModal, {
+        const modalClass = classNames(Styles.mobileModal, {
             [Styles.isOpen]: this.props.viewState.explorerPanelIsVisible && this.props.viewState.mobileView
         });
-        return <div className={modalClass}>
-                    {(this.props.viewState.explorerPanelIsVisible && this.props.viewState.mobileView) && <button type='button' className={Styles.closeModal} onClick={this.onClearMobileUI}>Done</button>}
+        const mobileView = this.props.viewState.mobileView;
+        
+        return (
+            <div className={modalClass}>
+                <div className={Styles.modalBg}>
+                    <div className={Styles.modalTop}>
+                        <If condition={this.props.viewState.explorerPanelIsVisible && mobileView}>
+                            <button type='button' className={Styles.doneButton} onClick={this.onClearMobileUI}>Done
+                            </button>
+                        </If>
+                        <button type='button'
+                                disabled={mobileView !== this.props.viewState.mobileViewOptions.preview}
+                                className={classNames(
+                                    Styles.backButton,
+                                    {[Styles.backButtonInactive]: mobileView !== this.props.viewState.mobileViewOptions.preview}
+                                )}
+                                onClick={this.goBack}>
+                            <Icon className={Styles.iconBack} glyph={Icon.GLYPHS.left}/>
+                        </button>
+                    </div>
+
                     {this.renderModalContent()}
-                </div>;
+                </div>
+            </div>
+        );
     }
 });
 module.exports = MobileModalWindow;
