@@ -3,7 +3,6 @@
 import defined from 'terriajs-cesium/Source/Core/defined';
 import knockout from 'terriajs-cesium/Source/ThirdParty/knockout';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
-import runLater from '../Core/runLater';
 
 const ObserveModelMixin = {
     componentWillMount() {
@@ -49,15 +48,9 @@ const ObserveModelMixin = {
 
                     if (!updateForced) {
                         updateForced = true;
-                        runLater(function() {
-                            updateForced = false;
-                            if (that.isMounted()) {
-                                that.forceUpdate();
-                            }
-                        }).otherwise(e => {
-                            console.error(`Render error in ${that.constructor.displayName || that.constructor.name}:\n${e.stack}`);
-                            throw e;
-                        });
+                        if (that.isMounted()) {
+                            that.forceUpdate();
+                        }
                     }
                 }
 
