@@ -71,7 +71,7 @@ const SidePanel = React.createClass({
                                searchText={searchState.locationSearchText} />
                     <div className={Styles.addData}>
                         <button type='button' onClick={this.onAddDataClicked} className={Styles.button}>
-                            <Icon glyph={Icon.GLYPHS.add}/>Add
+                            <Icon glyph={Icon.GLYPHS.add}/>Add data
                         </button>
                     </div>
                 </div>
@@ -95,7 +95,7 @@ const SidePanel = React.createClass({
                                     <li>Browse the Data Catalogue</li>
                                     <li>Load your own data onto the map</li>
                                 </ul>
-                                <p><Icon glyph={Icon.GLYPHS.bulb}/><strong>TIP:</strong> <em>All of your active data sets will be listed
+                                <p><Icon glyph={Icon.GLYPHS.bulb}/><strong>TIP:</strong> <em>All your active data sets will be listed
                                     here</em></p>
                             </div>
                         </Otherwise>
