@@ -64,28 +64,30 @@ const Concept = React.createClass({
         return (
             <li style={this.getColorStyle()}>
                 <If condition={concept.name}>
-                    <div className={classNames(Styles.header, {[Styles.hasChildren]: concept.hasChildren})}>
-                        <If condition={concept.hasChildren}>
-                            <button type='button'
-                                    onClick={this.toggleOpen}
-                                    style={this.getColorStyle()}
-                                    className={Styles.btn}
-                                    title='open variable selection'>
-                                    {concept.isOpen ? <Icon glyph={Icon.GLYPHS.opened}/> : <Icon glyph={Icon.GLYPHS.closed}/>}
-                            </button>
-                        </If>
-                        <If condition={concept.isSelectable}>
-                            <button type='button'
-                                    onClick={this.toggleActive}
-                                    style={this.getColorStyle()}
-                                    className={Styles.btn}
-                                    title='select variable'>
-                                    {(concept.isActive && this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.checkboxOn}/>}
-                                    {(!concept.isActive && this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.checkboxOff}/>}
-                                    {(concept.isActive && !this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.radioOn}/>}
-                                    {(!concept.isActive && !this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.radioOff}/>}
-                            </button>
-                        </If>
+                    <div className={classNames(Styles.header, {[Styles.hasChildren]: concept.hasChildren, [Styles.isSelectable]: concept.isSelectable})}>
+                        <div className={Styles.btnGroup}>
+                            <If condition={concept.hasChildren}>
+                                <button type='button'
+                                        onClick={this.toggleOpen}
+                                        style={this.getColorStyle()}
+                                        className={Styles.btnToggleOpen}
+                                        title='open variable selection'>
+                                        {concept.isOpen ? <Icon glyph={Icon.GLYPHS.showLess}/> : <Icon glyph={Icon.GLYPHS.showMore}/>}
+                                </button>
+                            </If>
+                            <If condition={concept.isSelectable}>
+                                <button type='button'
+                                        onClick={this.toggleActive}
+                                        style={this.getColorStyle()}
+                                        className={Styles.btnToggleActive}
+                                        title='select variable'>
+                                        {(concept.isActive && this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.checkboxOn}/>}
+                                        {(!concept.isActive && this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.checkboxOff}/>}
+                                        {(concept.isActive && !this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.radioOn}/>}
+                                        {(!concept.isActive && !this.props.allowMultiple) && <Icon style={this.getFillStyle()} glyph={Icon.GLYPHS.radioOff}/>}
+                                </button>
+                            </If>
+                        </div>
                         {concept.name}
                     </div>
                 </If>
